docs: 프로젝트 swift 파일 정리
