--- BankManagerConsoleApp/BankManagerConsoleApp/Model/Client.swift
@@ -1,13 +1,0 @@
-//
-//  Client.swift
-//  BankManagerConsoleApp
-//
-//  Created by Yongwoo Marco on 2021/08/02.
-//
-
-import Foundation
-
-struct Client {
-    let waitingNumber: Int
-    let bankingTask: BankingTask = BankingTask.allCases.randomElement() ?? .deposit
-}
